Major change at MainActivity.kt: The activity was immediately closed (via finish()) at onCreate if flag FLAG_ACTIVITY_BROUGHT_TO_FRONT was set in intent.flags. This was done to avoid multiple instances of this activity. But this caused the problem that the intents that were received at this aborted onCreate-call never got processed. For example an intent of a widget to show a specific list. To solve that intents were not processed the abortion by flag FLAG_ACTIVITY_BROUGHT_TO_FRONT was removed. To avoid that multiple instances get created, the MainActivity got android:launchMode="singleTask" in AndroidManifest.xml.
Other changes at MainActivity.kt:
- Doing GUI setup before the processing of extras to have the GUI ready

Clean-up at MainActivity.kt:
- TodoTasksFragment was created as currentFragment but not used at all. Removed currentFragment and TodoTasksFragment.kt.
- authAndGuiInit() was done in onCreate and in onResume. This is doubled because onResume gets called every time. Even if onCreate was called before. So removed authAndGuiInit() from onCreate.
- no longer un-checking entries of navigationBottomView because they never get checked.
- Moved code that was many times copied before/after showAllTasks or showTasksOfList into these methods to avoid copy&paste

TodoListWidget put extras into template-intent to send it to MainActivity. But this does not work. The extras need to be put into the fill-intents (which happens in TodoListWidgetViewsFactory.kt). To get this working, the templante-intent additionally needs the flags FLAG_UPDATE_CURRENT and FLAG_MUTABLE instead of FLAG_IMMUTABLE.

ExpandableTodoTaskAdapter: It has the feature to show list names of tasks if they are on a list and all tasks get shown. This feature was inactive (for unknown reason, possibly a mistake) and now was re-enabled.