Introduced reminder state at DB to know which reminder was shown and which needs to be shown.
Before this was derived from the reminder timestamp: Was it in the past, the reminder was assumed to be shown. Was it in the future, it was assumed that the reminder needs to be shwon. This did lead to the problem that if more than one reminder was set to one point in time, only one was shown and the others where overssen after the one has been displayed. Because as this happend the others were past and so assumed to be shown. This should fix issue #162.
Generally improved job handling.